Simplify zhpe_stats, first pass
John Byrne suggested simplifying the libzhpe_stats to capture
stat readings in records stored in a buffer that is periodically
written out.

John suggested buffer-based model that writes records when
subids start and stop.
He said to use non-temporal writes to the buffer to reduce cache pollution,
but this version does not use non-temporal writes.

John said don’t worry about breaking compatibility on the “sim” side as
long as there is a way forward there in the future.

Signed-off-by: Harumi Kuno <harumi.kuno@hpe.com>